haha
Just found jaye's using example scene as the map, but she just delete all the walls on the edge to make camera able to look through. "Well done" That I have to rebuild all the colliders in the edge, less to say that this asset's stairs doesn't even have colliders in original. So great on make my workload heavier. And since its a 3d scene with even some balcony on the opposite of the camera angle. I have to make a script to auto turn camera 180 degrees.